rename Toggle to ColorModeToggle
