+++ packages/docusaurus-theme-classic/src/theme/ColorModeToggle/styles.module.css
